New event handler. Changed in ClientApp. Move menu gui to menu_gui. Auto cleanup run.

+++ src/cc/util/math/Tuple2d.java
@@ -16,7 +16,7 @@ import java.lang.Math;
 
 
 /**
- * A generic 2-element tuple that is represented by double-precision  
+ * A generic 2-element tuple that is represented by double-precision
  * floating point x,y coordinates.
  *
  */
@@ -75,8 +75,8 @@ public abstract class Tuple2d implements java.io.Serializable, Cloneable {
      */
     public Tuple2d(Tuple2f t1)
     {
-	this.x = (double) t1.x;
-	this.y = (double) t1.y;
+	this.x = t1.x;
+	this.y = t1.y;
     }
 
     /**
@@ -102,7 +102,7 @@ public abstract class Tuple2d implements java.io.Serializable, Cloneable {
 
 
     /**
-     * Sets the value of this tuple from the 2 values specified in 
+     * Sets the value of this tuple from the 2 values specified in
      * the array.
      * @param t the array of length 2 containing xy in order
      */
@@ -122,7 +122,7 @@ public abstract class Tuple2d implements java.io.Serializable, Cloneable {
 	this.x = t1.x;
 	this.y = t1.y;
     }
- 
+
 
     /**
      * Sets the value of this tuple to the value of Tuple2f t1.
@@ -130,8 +130,8 @@ public abstract class Tuple2d implements java.io.Serializable, Cloneable {
      */
     public final void set(Tuple2f t1)
     {
-	this.x = (double) t1.x;
-	this.y = (double) t1.y;
+	this.x = t1.x;
+	this.y = t1.y;
     }
 
    /**
@@ -160,7 +160,7 @@ public abstract class Tuple2d implements java.io.Serializable, Cloneable {
     /**
      * Sets the value of this tuple to the vector sum of itself and tuple t1.
      * @param t1 the other tuple
-     */  
+     */
     public final void add(Tuple2d t1)
     {
         this.x += t1.x;
@@ -169,23 +169,23 @@ public abstract class Tuple2d implements java.io.Serializable, Cloneable {
 
 
     /**
-     * Sets the value of this tuple to the vector difference of 
-     * tuple t1 and t2 (this = t1 - t2).    
+     * Sets the value of this tuple to the vector difference of
+     * tuple t1 and t2 (this = t1 - t2).
      * @param t1 the first tuple
      * @param t2 the second tuple
-     */  
+     */
     public final void setSub(Tuple2d t1, Tuple2d t2)
     {
         this.x = t1.x - t2.x;
         this.y = t1.y - t2.y;
-    }  
+    }
 
 
     /**
      * Sets the value of this tuple to the vector difference of
      * itself and tuple t1 (this = this - t1).
      * @param t1 the other vector
-     */  
+     */
     public final void sub(Tuple2d t1)
     {
         this.x -= t1.x;
@@ -245,13 +245,13 @@ public abstract class Tuple2d implements java.io.Serializable, Cloneable {
      * @param s the scalar value
      * @param t1 the tuple to be multipled
      * @param t2 the tuple to be added
-     */  
+     */
     public final void scaleAdd(double s, Tuple2d t1, Tuple2d t2)
     {
-        this.x = s*t1.x + t2.x; 
-        this.y = s*t1.y + t2.y; 
-    } 
- 
+        this.x = s*t1.x + t2.x;
+        this.y = s*t1.y + t2.y;
+    }
+
 
     /**
      * Sets the value of this tuple to the scalar multiplication
@@ -274,7 +274,7 @@ public abstract class Tuple2d implements java.io.Serializable, Cloneable {
      * code value.  Two objects with different data members may return the
      * same hash value, although this is not likely.
      * @return the integer hash code value
-     */  
+     */
     @Override
     public int hashCode() {
 	long bits = 1L;
@@ -284,12 +284,12 @@ public abstract class Tuple2d implements java.io.Serializable, Cloneable {
     }
 
 
-   /**   
+   /**
      * Returns true if all of the data members of Tuple2d t1 are
      * equal to the corresponding data members in this Tuple2d.
      * @param t1  the vector with which the comparison is made
      * @return  true or false
-     */  
+     */
     public boolean equals(Tuple2d t1)
     {
         try {
@@ -299,13 +299,13 @@ public abstract class Tuple2d implements java.io.Serializable, Cloneable {
 
     }
 
-   /**   
+   /**
      * Returns true if the Object t1 is of type Tuple2d and all of the
      * data members of t1 are equal to the corresponding data members in
      * this Tuple2d.
      * @param t1  the object with which the comparison is made
      * @return  true or false
-     */  
+     */
     @Override
     public boolean equals(Object t1)
     {
@@ -320,11 +320,11 @@ public abstract class Tuple2d implements java.io.Serializable, Cloneable {
 
    /**
      * Returns true if the L-infinite distance between this tuple
-     * and tuple t1 is less than or equal to the epsilon parameter, 
+     * and tuple t1 is less than or equal to the epsilon parameter,
      * otherwise returns false.  The L-infinite
-     * distance is equal to MAX[abs(x1-x2), abs(y1-y2)]. 
+     * distance is equal to MAX[abs(x1-x2), abs(y1-y2)].
      * @param t1  the tuple to be compared to this tuple
-     * @param epsilon  the threshold value  
+     * @param epsilon  the threshold value
      * @return  true or false
      */
     public boolean epsilonEquals(Tuple2d t1, double epsilon)
@@ -346,7 +346,7 @@ public abstract class Tuple2d implements java.io.Serializable, Cloneable {
      * Returns a string that contains the values of this Tuple2d.
      * The form is (x,y).
      * @return the String representation
-     */  
+     */
    @Override
 public String toString()
    {
@@ -355,15 +355,15 @@ public String toString()
 
 
   /**
-    *  Clamps the tuple parameter to the range [low, high] and 
-    *  places the values into this tuple.  
+    *  Clamps the tuple parameter to the range [low, high] and
+    *  places the values into this tuple.
     *  @param min   the lowest value in the tuple after clamping
-    *  @param max  the highest value in the tuple after clamping 
+    *  @param max  the highest value in the tuple after clamping
     *  @param t   the source tuple, which will not be modified
     */
    public final void clamp(double min, double max, Tuple2d t)
    {
-        if( t.x > max ) { 
+        if( t.x > max ) {
           x = max;
         } else if( t.x < min ){
           x = min;
@@ -371,7 +371,7 @@ public String toString()
           x = t.x;
         }
 
-        if( t.y > max ) { 
+        if( t.y > max ) {
           y = max;
         } else if( t.y < min ){
           y = min;
@@ -382,62 +382,62 @@ public String toString()
    }
 
 
-  /** 
-    *  Clamps the minimum value of the tuple parameter to the min 
+  /**
+    *  Clamps the minimum value of the tuple parameter to the min
     *  parameter and places the values into this tuple.
-    *  @param min   the lowest value in the tuple after clamping 
+    *  @param min   the lowest value in the tuple after clamping
     *  @param t   the source tuple, which will not be modified
-    */   
-   public final void clampMin(double min, Tuple2d t) 
-   { 
-        if( t.x < min ) { 
+    */
+   public final void clampMin(double min, Tuple2d t)
+   {
+        if( t.x < min ) {
           x = min;
         } else {
           x = t.x;
         }
 
-        if( t.y < min ) { 
+        if( t.y < min ) {
           y = min;
         } else {
           y = t.y;
         }
 
-   } 
+   }
 
 
-  /**  
-    *  Clamps the maximum value of the tuple parameter to the max 
+  /**
+    *  Clamps the maximum value of the tuple parameter to the max
     *  parameter and places the values into this tuple.
-    *  @param max   the highest value in the tuple after clamping  
+    *  @param max   the highest value in the tuple after clamping
     *  @param t   the source tuple, which will not be modified
-    */    
-   public final void clampMax(double max, Tuple2d t)  
-   {  
-        if( t.x > max ) { 
+    */
+   public final void clampMax(double max, Tuple2d t)
+   {
+        if( t.x > max ) {
           x = max;
-        } else { 
+        } else {
           x = t.x;
         }
- 
+
         if( t.y > max ) {
           y = max;
         } else {
           y = t.y;
         }
 
-   } 
+   }
 
 
-  /**  
-    *  Sets each component of the tuple parameter to its absolute 
+  /**
+    *  Sets each component of the tuple parameter to its absolute
     *  value and places the modified values into this tuple.
     *  @param t   the source tuple, which will not be modified
-    */    
+    */
   public final void absolute(Tuple2d t)
   {
        x = Math.abs(t.x);
        y = Math.abs(t.y);
-  } 
+  }
 
 
 
@@ -453,7 +453,7 @@ public String toString()
         } else if( x < min ){
           x = min;
         }
- 
+
         if( y > max ) {
           y = max;
         } else if( y < min ){
@@ -462,24 +462,24 @@ public String toString()
 
    }
 
- 
+
   /**
     *  Clamps the minimum value of this tuple to the min parameter.
     *  @param min   the lowest value in this tuple after clamping
     */
    public final void clampMin(double min)
-   { 
+   {
       if( x < min ) x=min;
       if( y < min ) y=min;
-   } 
- 
- 
+   }
+
+
   /**
     *  Clamps the maximum value of this tuple to the max parameter.
     *  @param max   the highest value in the tuple after clamping
     */
    public final void clampMax(double max)
-   { 
+   {
       if( x > max ) x=max;
       if( y > max ) y=max;
    }
@@ -495,8 +495,8 @@ public String toString()
   }
 
 
-  /** 
-    *  Linearly interpolates between tuples t1 and t2 and places the 
+  /**
+    *  Linearly interpolates between tuples t1 and t2 and places the
     *  result into this tuple:  this = (1-alpha)*t1 + alpha*t2.
     *  @param t1  the first tuple
     *  @param t2  the second tuple
@@ -509,18 +509,18 @@ public String toString()
   }
 
 
-  /**  
-    *  Linearly interpolates between this tuple and tuple t1 and 
+  /**
+    *  Linearly interpolates between this tuple and tuple t1 and
     *  places the result into this tuple:  this = (1-alpha)*this + alpha*t1.
     *  @param t1  the first tuple
-    *  @param alpha  the alpha interpolation parameter  
-    */   
-  public final void interpolate(Tuple2d t1, double alpha) 
-  { 
+    *  @param alpha  the alpha interpolation parameter
+    */
+  public final void interpolate(Tuple2d t1, double alpha)
+  {
        this.x = (1-alpha)*this.x + alpha*t1.x;
        this.y = (1-alpha)*this.y + alpha*t1.y;
 
-  } 
+  }
 
     /**
      * Creates a new object of the same class as this object.
@@ -544,9 +544,9 @@ public String toString()
 
 	/**
 	 * Get the <i>x</i> coordinate.
-	 * 
+	 *
 	 * @return the <i>x</i> coordinate.
-	 * 
+	 *
 	 * @since vecmath 1.5
 	 */
 	public final double getX() {
@@ -556,9 +556,9 @@ public String toString()
 
 	/**
 	 * Set the <i>x</i> coordinate.
-	 * 
+	 *
 	 * @param x  value to <i>x</i> coordinate.
-	 * 
+	 *
 	 * @since vecmath 1.5
 	 */
 	public final void setX(double x) {
@@ -568,9 +568,9 @@ public String toString()
 
 	/**
 	 * Get the <i>y</i> coordinate.
-	 * 
+	 *
 	 * @return the <i>y</i> coordinate.
-	 * 
+	 *
 	 * @since vecmath 1.5
 	 */
 	public final double getY() {
@@ -580,9 +580,9 @@ public String toString()
 
 	/**
 	 * Set the <i>y</i> coordinate.
-	 * 
+	 *
 	 * @param y value to <i>y</i> coordinate.
-	 * 
+	 *
 	 * @since vecmath 1.5
 	 */
 	public final void setY(double y) {
