+++ src/cc/util/math/VecMathVector2d.java
@@ -16,7 +16,7 @@ import java.lang.Math;
 
 
 /**
- * A 2-element vector that is represented by double-precision floating 
+ * A 2-element vector that is represented by double-precision floating
  * point x,y coordinates.
  *
  */
@@ -69,7 +69,7 @@ public class VecMathVector2d extends Tuple2d implements java.io.Serializable {
     /**
      * Constructs and initializes a Vector2d from the specified Tuple2d.
      * @param t1 the Tuple2d containing the initialization x y data
-     */  
+     */
     public VecMathVector2d(Tuple2d t1)
     {
        super(t1);
@@ -79,7 +79,7 @@ public class VecMathVector2d extends Tuple2d implements java.io.Serializable {
     /**
      * Constructs and initializes a Vector2d from the specified Tuple2f.
      * @param t1 the Tuple2f containing the initialization x y data
-     */  
+     */
     public VecMathVector2d(Tuple2f t1)
     {
        super(t1);
@@ -105,19 +105,19 @@ public class VecMathVector2d extends Tuple2d implements java.io.Serializable {
     }
 
 
-    /**  
+    /**
      * Returns the length of this vector.
      * @return the length of this vector
-     */  
+     */
     public final double length()
     {
-        return (double) Math.sqrt(x*x + y*y);
+        return Math.sqrt(x*x + y*y);
     }
 
-    /**  
+    /**
      * Returns the squared length of this vector.
      * @return the squared length of this vector
-     */  
+     */
     public final double lengthSquared()
     {
         return (this.x*this.x + this.y*this.y);
@@ -126,25 +126,24 @@ public class VecMathVector2d extends Tuple2d implements java.io.Serializable {
     /**
      * Sets the value of this vector to the normalization of vector v1.
      * @param v1 the un-normalized vector
-     */  
+     */
     public final void normalize(VecMathVector2d v1)
     {
         double norm;
 
-        norm = (double) (1.0/Math.sqrt(v1.x*v1.x + v1.y*v1.y));
+        norm = (1.0/Math.sqrt(v1.x*v1.x + v1.y*v1.y));
         this.x = v1.x*norm;
         this.y = v1.y*norm;
     }
 
     /**
      * Normalizes this vector in place.
-     */  
+     */
     public final void normalize()
     {
         double norm;
 
-        norm = (double)
-               (1.0/Math.sqrt(this.x*this.x + this.y*this.y));
+        norm = (1.0/Math.sqrt(this.x*this.x + this.y*this.y));
         this.x *= norm;
         this.y *= norm;
     }
@@ -161,7 +160,7 @@ public class VecMathVector2d extends Tuple2d implements java.io.Serializable {
       double vDot = this.dot(v1) / ( this.length()*v1.length() );
       if( vDot < -1.0) vDot = -1.0;
       if( vDot >  1.0) vDot =  1.0;
-      return((double) (Math.acos( vDot )));
+      return((Math.acos( vDot )));
 
    }
