+++ src/cc/util/sound/SoundLoader.java
@@ -11,13 +11,13 @@ import org.lwjgl.openal.AL10;
 
 /**
  * Responsible for holding and playing the sounds used in the game.
- * 
+ *
  * @author Kevin Glass
  */
 public class SoundLoader {
 	/** The single instance of this class */
 	private static final SoundLoader store = new SoundLoader();
-	
+
 	/** True if sound effects are turned on */
 	private boolean sounds;
 	/** True if sound initialisation succeeded */
@@ -32,17 +32,17 @@ public class SoundLoader {
 	private int nextSource;
 	/** True if the sound system has been initialise */
 	private boolean inited = false;
-	
+
 	/**
 	 * Create a new sound store
 	 */
 	private SoundLoader() {
 		this.init();
 	}
-	
+
 	/**
-	 * Indicate whether sound effects should be played 
-	 * 
+	 * Indicate whether sound effects should be played
+	 *
 	 * @param sounds True if sound effects should be played
 	 */
 	public void setSoundsOn(boolean sounds) {
@@ -50,16 +50,16 @@ public class SoundLoader {
 			this.sounds = sounds;
 		}
 	}
-	
+
 	/**
 	 * Check if sound effects are currently enabled
-	 * 
+	 *
 	 * @return True if sound effects are currently enabled
 	 */
 	public boolean soundsOn() {
 		return sounds;
 	}
-	
+
 	/**
 	 * Initialise the sound effects stored. This must be called
 	 * before anything else will work
@@ -75,23 +75,23 @@ public class SoundLoader {
 			soundWorks = false;
 			sounds = false;
 		}
-		
+
 		if (soundWorks) {
 			sourceCount = 8;
 			sources = BufferUtils.createIntBuffer(8);
 			AL10.alGenSources(sources);
-			
+
 			if (AL10.alGetError() != AL10.AL_NO_ERROR) {
 				sounds = false;
 				soundWorks = false;
-			} 
+			}
 		}
 	}
 
 	/**
 	 * Play the specified buffer as a sound effect with the specified
 	 * pitch and gain.
-	 * 
+	 *
 	 * @param buffer The ID of the buffer to play
 	 * @param pitch The pitch to play at
 	 * @param gain The gain to play at
@@ -104,52 +104,52 @@ public class SoundLoader {
 					nextSource = 1;
 				}
 				AL10.alSourceStop(sources.get(nextSource));
-				
+
 				AL10.alSourcei(sources.get(nextSource), AL10.AL_BUFFER, buffer);
 				AL10.alSourcef(sources.get(nextSource), AL10.AL_PITCH, pitch);
-				AL10.alSourcef(sources.get(nextSource), AL10.AL_GAIN, gain); 
-				
-				AL10.alSourcePlay(sources.get(nextSource)); 
+				AL10.alSourcef(sources.get(nextSource), AL10.AL_GAIN, gain);
+
+				AL10.alSourcePlay(sources.get(nextSource));
 			}
 		}
 	}
-	
+
 	/**
 	 * Get the Sound based on a specified OGG file
-	 * 
+	 *
 	 * @param ref The reference to the OGG file in the classpath
 	 * @return The Sound read from the OGG file
 	 * @throws IOException Indicates a failure to load the OGG
 	 */
-	public Sound getOgg(String ref) throws IOException 
+	public Sound getOgg(String ref) throws IOException
 	{
 		if (!soundWorks) {
 			return new Sound(this, 0);
 		}
-		
+
 		if (!inited) {
 			throw new IllegalStateException("Can't load sounds until SoundLoader is init()");
 		}
 		int buffer = -1;
-		
+
 		if (loaded.get(ref) != null) {
-			buffer = ((Integer) loaded.get(ref)).intValue();
+			buffer = (loaded.get(ref)).intValue();
 		} else {
 			InputStream in = null;
 			try {
 				IntBuffer buf = BufferUtils.createIntBuffer(1);
-				
+
 				in = Thread.currentThread()
 					.getContextClassLoader().getResourceAsStream(ref);
-				
+
 				OggDecoder decoder = new OggDecoder();
 				OggData ogg = decoder.getData(in);
-				
+
 				AL10.alGenBuffers(buf);
 				AL10.alBufferData(buf.get(0), ogg.channels > 1 ? AL10.AL_FORMAT_STEREO16 : AL10.AL_FORMAT_MONO16, ogg.data, ogg.rate);
-				
+
 				loaded.put(ref,new Integer(buf.get(0)));
-				                     
+
 				buffer = buf.get(0);
 			} catch (Exception e) {
 				throw new IOException("Unable to load: "+ref);
@@ -162,17 +162,17 @@ public class SoundLoader {
 				}
 			}
 		}
-		
+
 		if (buffer == -1) {
 			throw new IOException("Unable to load: "+ref);
 		}
-		
+
 		return new Sound(this, buffer);
 	}
-	
+
 	/**
 	 * Get the single instance of this class
-	 * 
+	 *
 	 * @return The single instnace of this class
 	 */
 	public static SoundLoader get() {
@@ -183,6 +183,6 @@ public class SoundLoader {
     {
     	return sourceCount;
     }
-	
-	
+
+
 }
